Split analyzer tests between shared and app
