+++ app/src/test/java/ai/brokk/analyzer/CSharpAnalyzerTest.java
@@ -1,10 +1,7 @@
 package ai.brokk.analyzer;
 
 import static ai.brokk.testutil.AssertionHelperUtil.assertCodeEquals;
-import static ai.brokk.testutil.UsageFinderTestUtil.fileNamesFromHits;
-import static ai.brokk.testutil.UsageFinderTestUtil.newFinder;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
 import ai.brokk.AnalyzerUtil;
 import ai.brokk.IContextManager;
@@ -369,31 +366,6 @@ public final class CSharpAnalyzerTest {
         assertTrue(
                 handlerSkeleton.get().contains("public class GetTerminationRecordByIdHandler"),
                 "Handler skeleton should contain correct class name");
-    }
-
-    @Test
-    public void getUsesClassComprehensivePatternsTest() throws InterruptedException {
-        TestProject project = TestProject.createTestProject("testcode-cs", Languages.C_SHARP);
-        CSharpAnalyzer analyzer = new CSharpAnalyzer(project);
-
-        var finder = newFinder(project, analyzer);
-        var symbol = "BaseClass";
-        var either = finder.findUsages(symbol).toEither();
-
-        assumeFalse(either.hasErrorMessage(), "C# analyzer unavailable");
-
-        var hits = either.getUsages();
-        var files = fileNamesFromHits(hits);
-        assertTrue(
-                files.contains("ClassUsagePatterns.cs"),
-                "Expected comprehensive usage patterns in ClassUsagePatterns.cs; actual: " + files);
-
-        var classUsageHits = hits.stream()
-                .filter(h -> h.file().absPath().getFileName().toString().equals("ClassUsagePatterns.cs"))
-                .toList();
-        assertTrue(
-                classUsageHits.size() >= 2,
-                "Expected at least 2 different usage patterns, found: " + classUsageHits.size());
     }
 
     @Test
