+++ app/src/test/java/ai/brokk/analyzer/CppAnalyzerTest.java
@@ -2,8 +2,6 @@ package ai.brokk.analyzer;
 
 import static ai.brokk.testutil.AssertionHelperUtil.assertCodeContains;
 import static ai.brokk.testutil.AssertionHelperUtil.assertCodeEquals;
-import static ai.brokk.testutil.UsageFinderTestUtil.fileNamesFromHits;
-import static ai.brokk.testutil.UsageFinderTestUtil.newFinder;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assumptions.*;
 
@@ -1495,28 +1493,6 @@ public class CppAnalyzerTest {
         String skeleton = skeletons.get(fooDef);
         assertNotNull(skeleton, "Skeleton for definition should exist");
         assertCodeContains(skeleton, "{...}");
-    }
-
-    @Test
-    public void getUsesClassComprehensivePatternsTest() throws InterruptedException {
-        var finder = newFinder(testProject, analyzer);
-        var symbol = "BaseClass";
-        var either = finder.findUsages(symbol).toEither();
-
-        assumeFalse(either.hasErrorMessage(), "C++ analyzer unavailable");
-
-        var hits = either.getUsages();
-        var files = fileNamesFromHits(hits);
-        assertTrue(
-                files.contains("class_usage_patterns.cpp"),
-                "Expected comprehensive usage patterns in class_usage_patterns.cpp; actual: " + files);
-
-        var classUsageHits = hits.stream()
-                .filter(h -> h.file().absPath().getFileName().toString().equals("class_usage_patterns.cpp"))
-                .toList();
-        assertTrue(
-                classUsageHits.size() >= 2,
-                "Expected at least 2 different usage patterns, found: " + classUsageHits.size());
     }
 
     @Test
