+++ app/src/test/java/ai/brokk/analyzer/GoAnalyzerTest.java
@@ -1,10 +1,7 @@
 package ai.brokk.analyzer;
 
 import static ai.brokk.testutil.AssertionHelperUtil.assertCodeEquals;
-import static ai.brokk.testutil.UsageFinderTestUtil.fileNamesFromHits;
-import static ai.brokk.testutil.UsageFinderTestUtil.newFinder;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
 import ai.brokk.AnalyzerUtil;
 import ai.brokk.testutil.AnalyzerCreator;
@@ -1465,27 +1462,5 @@ public class GoAnalyzerTest {
                     """,
                     inlineAnalyzer.getSkeleton(prefs).orElseThrow());
         }
-    }
-
-    @Test
-    public void getUsesClassComprehensivePatternsTest() throws InterruptedException {
-        var finder = newFinder(testProject, analyzer);
-        var symbol = "main.BaseStruct";
-        var either = finder.findUsages(symbol).toEither();
-
-        assumeFalse(either.hasErrorMessage(), "Go analyzer unavailable");
-
-        var hits = either.getUsages();
-        var files = fileNamesFromHits(hits);
-        assertTrue(
-                files.contains("class_usage_patterns.go"),
-                "Expected comprehensive usage patterns in class_usage_patterns.go; actual: " + files);
-
-        var classUsageHits = hits.stream()
-                .filter(h -> h.file().absPath().getFileName().toString().equals("class_usage_patterns.go"))
-                .toList();
-        assertTrue(
-                classUsageHits.size() >= 2,
-                "Expected at least 2 different usage patterns, found: " + classUsageHits.size());
     }
 }
